feat(v0.2-P1): Silero VAD via rten → Top-1 34% → 37%, outdoor_noisy F1 +50%
After tract-onnx (op-compat) and ort (ABI mismatch) both failed, swap to
rten — Robert Knight's pure-Rust ONNX runtime — which has a working
Silero example we adapted directly. Requires .rten format (one-time
conversion via `pip install rten-convert`).

Direct A/B on identical 200-clip ESC-50 dataset:
- Top-1:           0.3400 → 0.3700 (+0.030)
- Macro-F1:        0.3721 → 0.3998 (+0.028)
- Cohen's kappa:   0.2247 → 0.2544 (+0.030)
- outdoor_noisy F1: 0.2128 → 0.3200 (+0.107) ← biggest win

Win is concentrated where heuristic VAD (energy + ZCR + spectral flatness)
over-fired voice on natural noise (rain/wind/sea). Silero correctly
suppresses these → outdoor_noisy `voice_ratio < 0.3` gate fires.

Remaining ceiling on multi_speaker_chat / near_silence is *not* a VAD
problem — it's the heuristic speaker-count stub and missing
energy-floor calibration (v0.3 work).

Signal Model preserved: rten loads from local .rten file, no network.
cargo deny continues to block all HTTP crates (reqwest/hyper/surf/ureq/awc).

Changes:
- perceptkit-audio: rten + rten-tensor deps under `silero-vad` feature
- perceptkit-audio: SileroVadExtractor (480-sample 30ms window, LSTM
  state threaded across chunks, mutex-wrapped Model for FeatureExtractor
  Sync requirement)
- perceptkit-audio: AudioProvider::with_defaults_silero(model_path)
- perceptkit-py: SceneEngine.from_dir_silero(scenes, model) staticmethod
  (gated by silero-vad feature on the py crate)
- scripts/bench_from_audio_dir.py: --silero flag
- silero_e2e tests: un-ignored (4 pass against models/silero_vad.rten)
- energy.rs: drop redundant `as f64` cast (clippy)
- .gitignore: real_bench*.jsonl

+++ crates/perceptkit-audio/src/extractors/energy.rs
@@ -126,7 +126,7 @@ impl FeatureExtractor for EnergyExtractor {
         let low_energy_ratio = if pcm.len() < sub_window || peak_v < 1e-9 {
             0.0
         } else {
-            let threshold = (peak_v as f64) * 0.10;
+            let threshold = peak_v * 0.10;
             let mut total = 0_usize;
             let mut low = 0_usize;
             for chunk in pcm.chunks(sub_window) {
